chore: Adjust `NEUTRAL` tones to prevent `Text` regressions
This subsequently makes `NEUTRAL` components more pronounced — specifically their text color, bust most significantly the `SOLID` variant's `background`.

This isn't a bad thing and has relatively low impact on `Notification` and `Badge` — in regards to `Button`, this is a good thing, because it actually lays the foundation to have a `disabled` state for `Button.SECONDARY` and `GHOST` eventually.

It also fixes the `hover` state for the `NEUTRAL` button.