Use "body" font family for <Text />
The Text component hardcodes the use of the "system" font stack. This unfortunately means that to globally use a different font family we have to override it in the theme like so:

```JS
const theme = {
  ...baseTheme,
  fonts: {
    ...baseTheme.fonts,
    system: '...other fonts...'
  }
}
```

That is really confusing! Suddenly, `fontFamily: 'system'` no longer shows system fonts :confused:

Instead, this PR changes the Text component to use the "body" font stack, which is aliased to "system" by default anyway so no behavior changes here. We can then adjust that font stack without anything being confusing! :tada: