Fix LongTooltipText story visibility…
by changing the position to `bottom`. Needed because we now set a `maxWidth` which results in part of the indeed _very long_ Tooltip text being displayed outside of the story viewport.

Not a real fix — a good reminder that we don't handle vertical collision detection with viewport boundaries (while we do for the horizontal axis).